third_party/blink/web_tests/external/wpt/webcodecs/audio-decoder.https.any.worker-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Unhandled rejection: Failed to execute 'isConfigSupported' on 'AudioDecoder': Invalid config; description is required for multi-channel Opus.
[FAIL] Test that AudioDecoder.configure() doesn't support config: Opus with more than two channels and without description
  Failed to execute 'configure' on 'AudioDecoder': Invalid config; description is required for multi-channel Opus.
[FAIL] Test that AudioDecoder.configure() doesn't support config: Opus with more than two channels and with a description that is too short
  assert_equals: expected "NotSupportedError" but got "OperationError"
[FAIL] Test that AudioDecoder.configure() doesn't support config: vorbis requires a description
  Failed to execute 'configure' on 'AudioDecoder': Invalid config; description is required.
[FAIL] Test that AudioDecoder.configure() doesn't support config: flac requires a description
  Failed to execute 'configure' on 'AudioDecoder': Invalid config; description is required.
Harness: the test ran to completion.
